Abstract Factory Pattern and Observer Pattern
SCENARIO:
In a store's ordering system, two types of orders can be made such as online order, shop order. Store is in ankara. A user would like to place an order for a 100-pound product from this system. But he wants to get information for two order types to decide the order. After placing orders for one of these order types with different shipping times, order notifications and payment types, the shipping status changes for both systems at different times. This change must be reported to the ordering system before it can be reported to the user. If the user chooses which system, the new cargo status should be transmitted to that system. At the same time, both types of orders can be selected from the user interface. So he decided to order from the store but he ordered an online order to guarantee his business against the possibility of not going to the store. In this case the cargo status should be reported to both ordering systems. But when the two orders are given at the same time, the total time increases.
Online orders are cargo within 30 min. Payment type is the method of transfer and the type of notification. 10 minutes while the order is being created, 10 minutes when the payment type and notification type are set according to this order, and 10 minutes when the shipping status is notified to the system. This time must be reported to the user. After 20 minutes the status of the cargo is determined and the notification is sent to the user via email. At the end of 30 minutes, the condition of the cargo is changed and the road is set as the output and the new notification to the user is sent again via email.
Store orders will be shipped within 2 hours. Store orders are made by phone. Payment type cash (payment at the door) and notifications are sent via SMS. After the payment type is set and transmitted to the user, the status of the cargo is determined and notified via sms. After this, the cargo departs within two hours and the new cargo situation is informed to the user again via sms.
When ordering from both the store and the online system, the user is added to the online system due to the order of the order in the store and the cargo goes out after 5 hours.

Prob 1)	For both types of orders, different notification types and different payment types have to be created.

Solution: Abstract Factory. An interface is created for notifications and payment types. Create an order factory. In this order factory, different notifications and different payment types are created according to order types.

Prob 2)	Our other problem is that the cargo situation changes after a certain period of time according to the order types.
Solution: Observer Pattern. The order factories are notified as the cargo status changes. Cargo subject, order factory observer.